--- libsrc/grabber/amlogic/ge2d_cmd.h
@@ -1,44 +1,0 @@
-/*
- * include/linux/amlogic/ge2d/ge2d_cmd.h
- *
- * Copyright (C) 2015 Amlogic, Inc. All rights reserved.
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
- * more details.
- *
-*/
-
-
-
-#ifndef _GE2D_CMD_H_
-#define _GE2D_CMD_H_
-
-#define	GE2D_STRETCHBLIT_NOALPHA_NOBLOCK    0x4708
-#define	GE2D_BLIT_NOALPHA_NOBLOCK           0x4707
-#define	GE2D_BLEND_NOBLOCK                  0x4706
-#define	GE2D_BLIT_NOBLOCK                   0x4705
-#define	GE2D_STRETCHBLIT_NOBLOCK            0x4704
-#define	GE2D_FILLRECTANGLE_NOBLOCK          0x4703
-
-
-#define	GE2D_STRETCHBLIT_NOALPHA            0x4702
-#define	GE2D_BLIT_NOALPHA                   0x4701
-#define	GE2D_BLEND                          0x4700
-#define	GE2D_BLIT                           0x46ff
-#define	GE2D_STRETCHBLIT                    0x46fe
-#define	GE2D_FILLRECTANGLE                  0x46fd
-#define	GE2D_SRCCOLORKEY                    0x46fc
-#define	GE2D_SET_COEF                       0x46fb
-#define	GE2D_CONFIG_EX                      0x46fa
-#define	GE2D_CONFIG                         0x46f9
-#define	GE2D_ANTIFLICKER_ENABLE             0x46f8
-
-#endif
-
